+++ src/extension.js
@@ -1,13 +1,14 @@
 /**
- * vs code plugin for highlighting TODOs and FIXMEs within your code
+ * vscode plugin for highlighting TODOs and FIXMEs within your code
+ * 
  * TODO:
+ * - [x]highlight customized text
+ * - [x]support customizing colors
  * - show corresponding message in status bar
  * - list all todos in command pannel
- * - highlight customized text
  * - add command to toggle the highlight
- * - https://github.com/Microsoft/vscode-eslint/blob/master/eslint/extension.ts
- * - https://code.visualstudio.com/docs/extensionAPI/extension-points#_contributesconfiguration
- * - Language Servers implementation
+ * - language servers implementation
+ * - show new release welcome message for every updates
  */
 
 var vscode = require('vscode');
@@ -15,24 +16,25 @@ var util = require('./util');
 
 function activate(context) {
 
-    var timeout = null;
+    let timeout = null;
 
-    var activeEditor = vscode.window.activeTextEditor;
+    let activeEditor = vscode.window.activeTextEditor;
 
-    var settings = vscode.workspace.getConfiguration('todohighlight');
-    let keywordsData = util.getKeywords(settings.get('keywords'));
-    let isCaseSensitive = settings.get('isCaseSensitive');
+    let settings = vscode.workspace.getConfiguration('todohighlight');
+    let isCaseSensitive = settings.get('isCaseSensitive', true);
 
-    var decorationTypes = {};
+    let keywordsData = util.getKeywords(settings.get('keywords'), isCaseSensitive);
+
+    let decorationTypes = {};
 
     Object.keys(keywordsData).forEach((v) => {
         if (!isCaseSensitive) {
-            v = v.toLowerCase()
+            v = v.toUpperCase()
         }
-        decorationTypes[v] = vscode.window.createTextEditorDecorationType(Object.assign({}, util.DEFAULT_STYLE, keywordsData[v]));
+        let mergedStyle = Object.assign({}, keywordsData[v]);
+        mergedStyle.overviewRulerColor = mergedStyle.backgroundColor;
+        decorationTypes[v] = vscode.window.createTextEditorDecorationType(mergedStyle);
     })
-    console.log('decorationTypes',JSON.stringify(decorationTypes));
-    
 
     let keywords = Object.keys(keywordsData).join('|');
     let pattern = new RegExp(keywords, 'g');
@@ -63,28 +65,28 @@ function activate(context) {
             return;
         }
 
-        var text = activeEditor.document.getText();
-        var mathes = {};
-        var match;
+        let text = activeEditor.document.getText();
+        let mathes = {};
+        let match;
         while (match = pattern.exec(text)) {
-            var startPos = activeEditor.document.positionAt(match.index);
-            var endPos = activeEditor.document.positionAt(match.index + match[0].length);
-            var decoration = {
+            let startPos = activeEditor.document.positionAt(match.index);
+            let endPos = activeEditor.document.positionAt(match.index + match[0].length);
+            let decoration = {
                 range: new vscode.Range(startPos, endPos),
-                //TODO: parse and show todo index
-                // hoverMessage: `TODO#${match.index}`
+                // TODO: parse and show corresponding content in popup panel when hover
+                // hoverMessage: ``
             };
 
-            var matchedValue = match[0];
+            let matchedValue = match[0];
             if (!isCaseSensitive) {
-                matchedValue = matchedValue.toLowerCase()
+                matchedValue = matchedValue.toUpperCase()
             }
             mathes[matchedValue] ? mathes[matchedValue].push(decoration) : (mathes[matchedValue] = [decoration]);
         }
 
         Object.keys(mathes).forEach((v) => {
             if (!isCaseSensitive) {
-                v = v.toLowerCase()
+                v = v.toUpperCase()
             }
             activeEditor.setDecorations(decorationTypes[v], mathes[v]);
         })