+++ src/util.js
@@ -1,5 +1,8 @@
 var vscode = require('vscode');
 
+let settings = vscode.workspace.getConfiguration('todohighlight');
+let defaultStyle = settings.get('defaultStyle');
+
 var DEFAULT_KEYWORDS = {
     "TODO:": {
         text: "TODO:",
@@ -24,11 +27,16 @@ var DEFAULT_STYLE = {
     overviewRulerLane: vscode.OverviewRulerLane.Right
 };
 
-function getKeywords(setting) {
+function getKeywords(setting, isCaseSensitive) {
     let result = DEFAULT_KEYWORDS;
-    setting.forEach((v, i) => {
+    setting.forEach((v) => {
         v = typeof v == 'string' ? { text: v } : v;
-        result[v.text] = Object.assign({}, DEFAULT_STYLE, v);
+        let text = v.text;
+        if (!text) return;//NOTE: in case of the text is empty or note set in configuration file
+        if (!isCaseSensitive) {
+            text = text.toUpperCase();
+        }
+        result[text] = Object.assign({}, DEFAULT_STYLE, defaultStyle, v);
     })
     return result;
 }