+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Change Log
-All notable changes to the "todo-highlight" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [0.1.0] - 2016-12-27
+- resolve #2, customizing colors support
+- resolve #3, customizing keywords support
+- case sensitive config now support in settings, see also #1
+- add MIT license
 
 ## [0.0.5] - 2016-12-27
 - Caseenable case-insensitive patterns , see [PR#1](https://github.com/wayou/vscode-todo-highlight/pull/1)